Delete deprecated extractImageDescriptionsAndUrls
