+++ package.json
@@ -13,7 +13,6 @@
     "test-extract-article-text": "tsx --trace-deprecation --env-file=.env ./src/consoleTests/testExtractArticleText.ts",
     "test-description": "tsx --trace-deprecation --env-file=.env ./src/consoleTests/testDescription.ts",
     "test-hero-images": "tsx --env-file=.env ./src/consoleTests/testHeroImages.ts",
-    "test-image-urls": "tsx --env-file=.env ./src/consoleTests/testImageUrls.ts",
     "test-post-new-links": "tsx --env-file=.env ./src/consoleTests/testPostNewLinks.ts",
     "test-post-test-links": "tsx --env-file=.env ./src/consoleTests/testPostTestLinks.ts",
     "test-summarize": "tsx --trace-deprecation --env-file=.env ./src/consoleTests/testSummarize.ts",
