+++ src/extractImageUrl/extractImageUrl.ts
